[NO ISSUE][CLUS] ACK Heartbeats from NCs
On HB from NC, respond with an ACK.  In the event a heartbeat is
rec'd from an unknown NC, attempt to respond with an exception to
give the chance for the NC to handle.  On receipt of an exception
on ACK indicating the node is not known to the CC, force a reconnect.

Change-Id: I7be64fd8c550a697729c7fcf8783beae95043cd7
Reviewed-on: https://asterix-gerrit.ics.uci.edu/2881
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Contrib: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Murtadha Hubail <mhubail@apache.org>