[ASTERIXDB-2567][RT] Fix serializing nullable record fields
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
When serializing a record, a field should be checked if it's nullable.
The value to be serialized should be checked if it's null or value to
serialize it correctly. Exceptions could happen when the record is in
an array if record fields are not checked.

Change-Id: I42105965e69c60bc4b123337296476b4d75f0ad8
Reviewed-on: https://asterix-gerrit.ics.uci.edu/3398
Contrib: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Murtadha Hubail <mhubail@apache.org>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>