[ASTERIXDB-2933][COMP][EXT] Pushdowns Part1: Expected Schema Node
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
We want to have a more aggressive pushdown for Parquet. To do so,
we need to compute the expected schema from the expressions in the
query plan. This change introduces the data structure that holds the
expected schema information.

Note:
Moved from asterix-om to asterix-algebra as the expected schema should
not be exposed outside the compiler.

Change-Id: I4b5aef414050a518fe230326e92ed66734b7b1d2
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/12765
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Wael Alkowaileet <wael.y.k@gmail.com>
Reviewed-by: Dmitry Lychagin <dmitry.lychagin@couchbase.com>