[ASTERIXDB-2960][IDX] LOJ array-index INLJ probe PK fix
- user model changes: no
- storage format changes: no
- interface changes: no

Type errors were being thrown for LOJ on subqueries involving UNNESTed
values for array index INLJ. This was due to a bug where we would try to
perform a DISTINCT on probe PK vars that weren't present. If we cannot
infer any probe entries from the context, then insert a UUID on top of
the probe to allow the DISTINCT to work as intended.

Change-Id: I1eccce55b8d4d2fdb47f678d89fcf957f0ab2ebe
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/13183
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Dmitry Lychagin <dmitry.lychagin@couchbase.com>