[NO ISSUE][COMP] Better reporting of error msg for index creation
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
When creating an index, if the type of the indexed field
could not be found, report the field name. For array index,
if there are no projections (no field names) and only
UNNESTs, report that the type of the elmenets of the
unnested field could not be found.

Change-Id: I6f2aea4a745ff84ceebcc3989fef5fd2b7f4d9e5
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/16524
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>
Reviewed-by: Dmitry Lychagin <dmitry.lychagin@couchbase.com>