[ASTERIXDB-3099][COMP] Enable intersection of array indexes
- user model changes: no
- storage format changes: no
- interface changes: no

Details:

A small fix to allow array indexes to be intersected with other array
indexes when the same array is shared. This fix does **not** enable
array indexes to be intersected with atomic/column indexes, nor does
it allow array indexes to be intersected with other array indexes when
their arrays are not shared (these would involve a more significant
design change).

Change-Id: I0aa1c28298cb898fc512e908aa12fd8a896375a4
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17318
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17354
Reviewed-by: Michael Blow <mblow@apache.org>