[ASTERIXDB-3313][COMP] Defuse profile when not profiling
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
The logical2physical map is time consuming to produce, so it should
not be computed unless it's needed for a profile. Likewise keeping
the last plan in memory is wasteful and useless unless a profile is
needed. Lastly in the pipeline assembler, look at the profile flag
first instead of the instanceof check to save time.

Change-Id: Ia9a84eedbcb33c29f03155a8605bb82af372f1f3
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/17962
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Ian Maxon <imaxon@uci.edu>
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>