[ASTERIXDB-3472][HYR] Use CleanupUtils.fail/close in ProbeAndJoinActivityNode
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
In the close() of the hash-join ProbeAndJoinActivityNode,
use CleanupUtils.fail()/close() when closing the next writer.

The reason is if all the build partitions are spilled, pushing
the frames from the ProbeAndJoinActivityNode to the next writer
will happen in the close(), not nextFrame().

In such cases, any exception that happens in the next writer
is caught in ProbeAndJoinActivityNode.close(). After that,
it calls fail()/close() on the next writer. Those fail()/close()
should be done using CleanupUtils so that the original caught
exception is not lost.

Ext-ref: MB-62949
Change-Id: I1cdee5ed457b72b4f6b3731fba3542898dec9bf3
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18707
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Michael Blow <mblow@apache.org>