[ASTERIXDB-3477,3474][RT] Profiling fixes and improvements
- user model changes: yes
- storage format changes: no
- interface changes: no

Details:

The operator id was being improperly used as the key when updating
the min and max time. It needs to include the activity id to clearly
differentiate between the different activities of an operator.

Exchange time was previously being reported as the difference between
the open and close time of the connector, which is not right. Calculating
the actual time a connector is taking will require careful consideration
of each connector type and usage, such as 1-N vs N-M and pipelined vs.
materializing. Until then it is better to simply omit it, because it is
not typically a predominant factor in query time  If it is, the sampled
partition profile still exists and will be useful to find those cases.

Previously, operators that were the last before a non 1-1 exchange would
not have time or cardinalities reported. This is now fixed.

All operators and connectors now report total cardinalities aggregated
across all partitions.

Ext-ref:MB-63566
Change-Id: I172f0044112777aec7200a0c6ae906711fcdc5f2
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18905
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>