[ASTERIXDB-3504][STO] Error in calculating if Page 0 is full for Column Datasets
- user model changes: no
- storage format changes: no
- interface changes: no

Details:
While checking if page 0 is full we were considering the split key from
the previous tuple and not the current tuple. In case the split key from
the current tuple is considerably larger, we might mark the page as not
full even when adding the current tuple to the page will overflow it.

Ext-ref: MB-63590
Change-Id: I5c53e573e8d2333161601a82af16f00120122579
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/18925
Reviewed-by: Wail Alkowaileet <wael.y.k@gmail.com>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>