[ASTERIXDB-3628][RT] Correct subplan profiles
- user model changes: no
- storage format changes: no
- interface changes: no

Details:

Currently subplan timing information is calculated during close().
This can lead to inaccuracies in the timing, especially when the
subplan is near the end of a task pipeline, because other
operator's times are calculated after close, during deinitialize().

Correct this by adding the capability in the MetaOperator to invoke a
method on each ProfiledPushRuntime SubplanPushRuntime to calculate the
timings, so it can be invoked at the proper time as to retrieve the
correct time from its downstream operators.

Ext-ref: MB-67499

Change-Id: I4d38710626778ef52abb1cd4fe1e0b26e0d74bec
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/20069
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Ali Alsuliman <ali.al.solaiman@gmail.com>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>