[ASTERIXDB-3659][EXT]: delegate assume role auth to AWS SDK
Details:
when using assume role authentication for AWS SDK,
we are using StsAssumeRoleCredentialsProvider, this
provider will refresh the assume role credentials
based on the provided configuration.

However, as this provider stays open and refreshes
itself, it requires the following:
- Closing this provider manually when done using it.
- STS client to talk to the STS service to assume role.
  - Which in turn requires manual closing when done.

The method for building clients was changed to return
CloseableAwsClients which contains:
- Consuming client
  This is the target client we are building, e.g., S3 client
  or Glue client.
- STS client
  This is the STS client used for talking with
  STS service to assume the role.
- Credentials Provider
  The credentials provider used for the consume client.

The above maintains a refrence for all clients and
credential providers so they all can be closed after
we are done using them.

Ext-ref: MB-68987
Change-Id: I6a3c755f94d377b443f21594443fac830875610e
Reviewed-on: https://asterix-gerrit.ics.uci.edu/c/asterixdb/+/20493
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Reviewed-by: Hussain Towaileb <hussainht@gmail.com>
Reviewed-by: Murtadha Hubail <mhubail@apache.org>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>